pyproject.toml
[tool.poetry]
name = "faker_companies"
version = "0.1.0"
description = "companies API for UI development"
authors = ["Wong Jiang Fung <jeff@leadbook.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pytz = "2021.3"
Django = "4.0.2"
django-configurations = "2.3.1"
gunicorn = "20.1.0"
newrelic = "7.2.4.171"
psycopg2-binary = "2.9.3"
dj-database-url = "0.5.0"
django-model-utils = "4.2.0"
django-unique-upload = "0.2.1"
djangorestframework = "3.13.1"
Markdown = "3.3.6"
django-filter = "21.1"
ipdb = "0.13.9"
ipython = "8.0.1"
mkdocs = "1.2.3"
flake8 = "4.0.1"
mock = "4.0.3"
factory-boy = "3.2.1"
django-nose = "1.4.7"
nose-progressive = "1.5.2"
coverage = "6.2"
django-storages = "1.12.3"
boto3 = "1.20.37"
Faker = "^13.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
